fix Dublication bug in User services and incrise jwt token time till 5 hours and created a new funtion in User repo . Tip make your own logic for accessing data in Db . next use email for storing tokne in redis

+++ FinalYearProject/src/main/java/com/FinalYearProject/FinalYearProject/Service/UserService.java
@@ -206,15 +206,14 @@ public class UserService {
     public Boolean verifyTokenAndOTP(String token, int otp) {
         try {
             Conformation conformation=redisService.get(token,Conformation.class);
-            if ( conformation !=null && otp == conformation.getOtp()) {
-                User user = conformation.getUser();
-                user.setExpired(false);
-                user.setLocked(false);
-                user.setIs_enable(true);
-                user.setLocked(false);
-                user.setExpired(false);
-                userRepository.save(user);
-                redisService.delete(token);  //delete conformation after verifying token and otp
+            if ( conformation !=null &&
+                    otp == conformation.getOtp() &&
+                    userRepository.existsByEmail(conformation.getUser().getEmail())
+            ) {
+                userRepository.updateIsEnableLockedExpiredToTrue(conformation.getUser().getEmail());
+
+                redisService.delete(token);//delete conformation after verifying token and otp
+
                 return Boolean.TRUE;
             }
             else {
